Made separate directories for error classes for better organization.
