+++ src/messages/errors/mod_error/ModError.java
@@ -1,0 +1,18 @@
+package messages.errors.mod_error;
+
+import messages.errors.Error;
+import utilities.PrettyPrint;
+
+public class ModError extends Error {
+
+    public ModError() {}
+
+    @Override
+    public String header() {
+        if(fileName != null) {
+            return PrettyPrint.YELLOW + "Modifier error detected in "
+                    + PrettyPrint.RESET + fileName() + "\n";
+        }
+        return PrettyPrint.YELLOW + "Modifier error detected!\n\n" + PrettyPrint.RESET;
+    }
+}
