+++ src/messages/errors/mod_error/ModErrorFactory.java
@@ -1,0 +1,8 @@
+package messages.errors.mod_error;
+
+import messages.errors.Error;
+import messages.errors.ErrorFactory;
+
+public class ModErrorFactory implements ErrorFactory {
+    public Error createError() { return new ModError(); }
+}
