+++ src/messages/errors/syntax_error/SyntaxError.java
@@ -1,0 +1,10 @@
+package messages.errors.syntax_error;
+
+import messages.errors.Error;
+
+public class SyntaxError extends Error {
+    public SyntaxError() {}
+    @Override
+    public String header() { return ""; }
+
+}
