+++ src/messages/errors/syntax_error/SyntaxErrorFactory.java
@@ -1,4 +1,7 @@
-package messages.errors;
+package messages.errors.syntax_error;
+
+import messages.errors.Error;
+import messages.errors.ErrorFactory;
 
 public class SyntaxErrorFactory implements ErrorFactory {
     public Error createError() { return new SyntaxError(); }
