+++ src/messages/errors/type_error/TypeErrorFactory.java
@@ -1,0 +1,8 @@
+package messages.errors.type_error;
+
+import messages.errors.Error;
+import messages.errors.ErrorFactory;
+
+public class TypeErrorFactory implements ErrorFactory {
+    public Error createError() { return new TypeError(); }
+}
